WAVE-3952 Add 'updatehost' parameter to operators
Changes the 'update' behaviour to use `#pragma acc upddate self` instead of `#pragma acc exit data copyout` to prevent the implicit deletion of data from the device.

Adds an 'updatehost' parameter to operators to control whether or not this copy gets executed, to allow intermediate operators to elide unnecessary copies.