WAVE-5033 Allow explicit suppression of loop fusion per-Eq to avoid overrunning L1 instruction cache in generated code (#128)
On sufficiently complex operators with many terms, Devito may over-fuse the clusters / loops to the point where the generated binary code for the inner body of a loop exceeds the size of the L1 instruction cache on the destination device. If this happens, then (depending on the L1 filtrate of the device) you can see a performance hit of between 50% and 400%.

Properly fixing this would involve tight integration with the binary code generation in the C compiler to detect when a code block is too large and then fissioning the loop and re-generating code, but a relatively easy end-user fix is allowing the user to specify a set of dimensions along which a given Eq should not be fused with other clusters. This does result in a reduction of optimisation opportunities (notably, CSE won't apply across these clusters to the same extent) but the overall wall clock runtime is dramatically reduced over the L1-thrashing case.

Usage is something along the lines of:

```python
# gather the space dimensions and don't fuse on them
dims = [d for d in lhs.grid.dims if d.is_Space]
x = Eq(lhs, rhs, suppress_fusion_dims=dims)
```

Passing a TimeDimension in will probably result in an unresolvable operator for any case where multiple Eqs refer to each other's output over time.