WAVE-5060 Add topological sort to the CSE pass (#129)
In cases where the user-provided expressions input to the CSE pass declare one or more local Symbols (usually a Constant or a Scalar, as in the angle-dependent density case in WAVE-4948), the CSE pass can extract temporary scalar symbols that depend on the local symbol and then incorrectly order them before the expression defining the local symbol, causing compile-time errors.

This commit adds a topological sort step after extracting the temporary symbols, ensuring that the combined set of input and generated expressions after the CSE pass remain in the correct order. The only dependencies considered by this toposort are the lhs of each expression in the list after the CSE temporaries are generated.